refactor(cli): split readonly.tsx into per-domain report modules + barrel
src/core/cli/ui/readonly.tsx (2,809 lines) → 11 per-domain modules under
src/core/cli/ui/reports/ (format, command-meta, runtime, settings, tasks, workflows,
plugins, packages, search, schedule, trash) + a readonly.ts barrel (export * from each).
Pure mechanical relocation per APPENDIX-cohesion.md — no behavior change, no DTO retyping,
every export name preserved. The ~30 dynamic import('.../cli/ui/readonly') call sites and
tests/cli/readonly-ui.test.tsx (imports all 34 components by name) resolve through the
barrel unchanged. doctor.tsx + doctor-repair.tsx now import the shared plural() from
reports/format.

Deviation from the appendix: doctor-repair's local valueText is NOT a true dup of
format's (format stringifies objects/arrays; doctor-repair returns the fallback) — kept
local to preserve rendering. Only plural() was de-duped there.

Verified: typecheck clean; tests/cli 252/0 (--isolate); readonly-ui 24/0; full binary
build (3 targets) compiles the split through src/core/cli.ts.
